sdig: add rudimentary EDE output
